+++ Cargo.toml
@@ -23,13 +23,13 @@ serde_json="1.0.120"
 tracing           ="0.1.40"
 tracing-subscriber={ version="0.3.18", features=["env-filter"] }
 
-notary-server={ git="https://github.com/tlsnotary/tlsn.git", tag="v0.1.0-alpha.7" }
 tlsn-verifier={ git="https://github.com/tlsnotary/tlsn.git", tag="v0.1.0-alpha.7" }
 tlsn-prover  ={ git="https://github.com/tlsnotary/tlsn.git", tag="v0.1.0-alpha.7" }
 tlsn-common  ={ git="https://github.com/tlsnotary/tlsn.git", tag="v0.1.0-alpha.7" }
 tlsn-core    ={ git="https://github.com/tlsnotary/tlsn.git", tag="v0.1.0-alpha.7" }
-tls-client   ={ git="https://github.com/tlsnotary/tlsn.git", tag="v0.1.0-alpha.7", package="tlsn-tls-client" }
-tlsn-utils   ={ git="https://github.com/tlsnotary/tlsn-utils", rev="e7b2db6" }
+tlsn-formats ={ git="https://github.com/tlsnotary/tlsn.git", tag="v0.1.0-alpha.7" }
+tlsn-utils   ={ git="https://github.com/tlsnotary/tlsn-utils", rev="e7b2db6", package="tlsn-utils" }
+spansy       ={ git="https://github.com/tlsnotary/tlsn-utils", rev="e7b2db6", package="spansy" }
 
 tls-client2       ={ git="https://github.com/pluto/tls-origo-legacy", package="tls-client" }
 tls-core          ={ git="https://github.com/pluto/tls-origo-legacy", package="tls-core" }
